requirements.txt
crewai>=0.28.8
crewai[tools]>=0.28.8
opencv-python>=4.8.0
numpy>=1.24.0
requests>=2.31.0
pydantic>=2.0.0
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
python-dotenv>=1.0.0
pathlib
dataclasses